Update README.md
Welcome to TRUST AI ecosystem. An ecosystem built on the Binance Blockchain. TRUST AI is an Ecosystem built on the Binance Blockchain with integrated AI. Trust AI is focused on helping developers, companies and even individuals with zero knowledge of coding or any other protocol to be able to create smart contracts, security reporting within minutes and also with Storage system Management